# docs/eci-files/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Neta-Resume API",
  "description": "Read-only resume aggregate for Indian legislators. Every fact carries provenance.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/eci-files/timeline": {
   "get": {
    "tags": [
     "eci-files"
    ],
    "summary": "Timeline",
    "description": "Entries matching the filter (date ascending, then id), plus topic/people/lane/state facets and\nchecked/unchecked counts scoped to that same filtered set. `fields=compact` returns just enough\nper entry to draw the lane timeline's dots (no citations). An unknown `state` slug returns an empty\n`entries` list, the same as an unknown `person`.",
    "operationId": "timeline_eci_files_timeline_get",
    "parameters": [
     {
      "name": "topic",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Topic"
      }
     },
     {
      "name": "person",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Person"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "lane",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Lane"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "fields",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fields"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/EciTimeline"
          },
          {
           "$ref": "#/components/schemas/EciTimelineCompact"
          }
         ],
         "title": "Response Timeline Eci Files Timeline Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/eci-files/states/{slug}": {
   "get": {
    "tags": [
     "eci-files"
    ],
    "summary": "State Page",
    "description": "One State/UT's summary and notes. The web fetches its entries separately, through the\ntimeline's `state=` filter.",
    "operationId": "state_page_eci_files_states__slug__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EciStatePage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/eci-files/people/{slug}": {
   "get": {
    "tags": [
     "eci-files"
    ],
    "summary": "Person Page",
    "description": "One person's profile (if linked) plus their full timeline.",
    "operationId": "person_page_eci_files_people__slug__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EciPersonPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/eci-files/entries/{entry_id}": {
   "get": {
    "tags": [
     "eci-files"
    ],
    "summary": "Get Entry",
    "operationId": "get_entry_eci_files_entries__entry_id__get",
    "parameters": [
     {
      "name": "entry_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entry Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EciEntry"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EciCheckCounts": {
    "properties": {
     "checked": {
      "type": "integer",
      "title": "Checked"
     },
     "unchecked": {
      "type": "integer",
      "title": "Unchecked"
     }
    },
    "type": "object",
    "required": [
     "checked",
     "unchecked"
    ],
    "title": "EciCheckCounts"
   },
   "EciCitation": {
    "properties": {
     "position": {
      "type": "integer",
      "title": "Position"
     },
     "url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url"
     },
     "publisher": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Publisher"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "published": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Published"
     },
     "tier": {
      "type": "integer",
      "title": "Tier"
     },
     "archive_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archive Url"
     },
     "quote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quote"
     }
    },
    "type": "object",
    "required": [
     "position",
     "tier"
    ],
    "title": "EciCitation"
   },
...
   "EciEntry": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "area": {
      "type": "string",
      "title": "Area"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "date_precision": {
      "type": "string",
      "title": "Date Precision"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "summary": {
      "type": "string",
      "title": "Summary"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "lane": {
      "type": "string",
      "title": "Lane"
     },
     "attributed_to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Attributed To"
     },
     "topics": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Topics",
      "default": []
     },
     "states": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "States",
      "default": []
     },
     "figures": {
      "items": {},
      "type": "array",
      "title": "Figures"
     },
     "details": {
      "additionalProperties": true,
      "type": "object",
      "title": "Details"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "check_status": {
      "type": "string",
      "title": "Check Status"
     },
     "people": {
      "items": {
       "$ref": "#/components/schemas/EciPersonRef"
      },
      "type": "array",
      "title": "People",
      "default": []
     },
     "response_to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Response To"
     },
     "responses": {
      "items": {
       "$ref": "#/components/schemas/EciResponseRef"
      },
      "type": "array",
      "title": "Responses",
      "default": []
     },
     "citations": {
      "items": {
       "$ref": "#/components/schemas/EciCitation"
      },
      "type": "array",
      "title": "Citations",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "area",
     "kind",
     "date_precision",
     "title",
     "summary",
     "status",
     "lane",
     "check_status"
    ],
    "title": "EciEntry"
   },
   "EciEntryCompact": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "date_precision": {
      "type": "string",
      "title": "Date Precision"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "lane": {
      "type": "string",
      "title": "Lane"
     },
     "check_status": {
      "type": "string",
      "title": "Check Status"
     },
     "people": {
      "items": {
       "$ref": "#/components/schemas/EciPersonRef"
      },
      "type": "array",
      "title": "People",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "date_precision",
     "title",
     "status",
     "lane",
     "check_status"
    ],
    "title": "EciEntryCompact",
    "description": "`fields=compact` on `/eci-files/timeline` \u2014 just enough to draw a dot, no citations."
   },
   "EciEntryRef": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "date_precision": {
      "type": "string",
      "title": "Date Precision"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "date_precision",
     "status"
    ],
    "title": "EciEntryRef",
    "description": "A compact, labelled reference to an entry cited by a selection, regime or departure."
   },
...
   "EciLaneCount": {
    "properties": {
     "lane": {
      "type": "string",
      "title": "Lane"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     }
    },
    "type": "object",
    "required": [
     "lane",
     "count"
    ],
    "title": "EciLaneCount"
   },
...
   "EciPersonCount": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name",
     "count"
    ],
    "title": "EciPersonCount"
   },
   "EciPersonDetail": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "profile": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciEntry"
       },
       {
        "type": "null"
       }
      ]
     },
     "group": {
      "type": "string",
      "title": "Group"
     },
     "current": {
      "type": "boolean",
      "title": "Current"
     },
     "role": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     },
     "service": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Service"
     },
     "tenure": {
      "items": {},
      "type": "array",
      "title": "Tenure"
     },
     "status_counts": {
      "$ref": "#/components/schemas/EciStatusCounts"
     },
     "photo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciPhoto"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name",
     "group",
     "current"
    ],
    "title": "EciPersonDetail"
   },
   "EciPersonPage": {
    "properties": {
     "person": {
      "$ref": "#/components/schemas/EciPersonDetail"
     },
     "entries": {
      "items": {
       "$ref": "#/components/schemas/EciEntry"
      },
      "type": "array",
      "title": "Entries"
     },
     "selections": {
      "items": {
       "$ref": "#/components/schemas/EciSelection"
      },
      "type": "array",
      "title": "Selections"
     },
     "entries_index": {
      "items": {
       "$ref": "#/components/schemas/EciEntryRef"
      },
      "type": "array",
      "title": "Entries Index"
     }
    },
    "type": "object",
    "required": [
     "person",
     "entries",
     "selections",
     "entries_index"
    ],
    "title": "EciPersonPage"
   },
   "EciPersonRef": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name"
    ],
    "title": "EciPersonRef"
   },
...
   "EciPhoto": {
    "properties": {
     "url": {
      "type": "string",
      "title": "Url"
     },
     "source_page": {
      "type": "string",
      "title": "Source Page"
     },
     "attribution": {
      "type": "string",
      "title": "Attribution"
     },
     "licence": {
      "type": "string",
      "title": "Licence"
     },
     "licence_url": {
      "type": "string",
      "title": "Licence Url"
     },
     "licence_review": {
      "type": "string",
      "title": "Licence Review"
     },
     "original_publisher": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Original Publisher"
     },
     "caption": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Caption"
     },
     "photo_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo Date"
     }
    },
    "type": "object",
    "required": [
     "url",
     "source_page",
     "attribution",
     "licence",
     "licence_url",
     "licence_review"
    ],
    "title": "EciPhoto"
   },
   "EciRegionSummary": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "phase": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase"
     },
     "exercise": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exercise"
     },
     "has_figures": {
      "type": "boolean",
      "title": "Has Figures"
     },
     "entry_count": {
      "type": "integer",
      "title": "Entry Count"
     },
     "stages": {
      "items": {
       "$ref": "#/components/schemas/EciStageValue"
      },
      "type": "array",
      "title": "Stages",
      "default": []
     },
     "metrics": {
      "$ref": "#/components/schemas/EciStateMetrics"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name",
     "code",
     "kind",
     "has_figures",
     "entry_count",
     "metrics"
    ],
    "title": "EciRegionSummary"
   },
   "EciResponseRef": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title"
    ],
    "title": "EciResponseRef",
    "description": "One entry that answers another (`response_to` points back at the charge it responds to)."
   },
   "EciSelection": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "date_precision": {
      "type": "string",
      "title": "Date Precision"
     },
     "date_meaning": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date Meaning"
     },
     "regime": {
      "type": "string",
      "title": "Regime"
     },
     "method": {
      "type": "string",
      "title": "Method"
     },
     "appointed": {
      "items": {
       "$ref": "#/components/schemas/EciSelectionAppointee"
      },
      "type": "array",
      "title": "Appointed"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/EciSelectionMember"
      },
      "type": "array",
      "title": "Members"
     },
     "search": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciSelectionSearch"
       },
       {
        "type": "null"
       }
      ]
     },
     "dissent": {
      "items": {
       "$ref": "#/components/schemas/EciSelectionDissent"
      },
      "type": "array",
      "title": "Dissent"
     },
     "entry_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entry Ids"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "id",
     "date",
     "date_precision",
     "regime",
     "method",
     "appointed",
     "members",
     "dissent",
     "entry_ids"
    ],
    "title": "EciSelection"
   },
   "EciSelectionAppointee": {
    "properties": {
     "person_slug": {
      "type": "string",
      "title": "Person Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "office": {
      "type": "string",
      "title": "Office"
     },
     "took_charge": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Took Charge"
     },
     "replaced": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Replaced"
     },
     "photo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciPhoto"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "person_slug",
     "name",
     "office"
    ],
    "title": "EciSelectionAppointee"
   },
   "EciSelectionDissent": {
    "properties": {
     "person_slug": {
      "type": "string",
      "title": "Person Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "summary": {
      "type": "string",
      "title": "Summary"
     },
     "note_public": {
      "type": "boolean",
      "title": "Note Public"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "entry_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entry Ids"
     },
     "response_entry_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Response Entry Ids",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "person_slug",
     "name",
     "summary",
     "note_public",
     "status",
     "entry_ids"
    ],
    "title": "EciSelectionDissent"
   },
   "EciSelectionMember": {
    "properties": {
     "person_slug": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Person Slug"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "part": {
      "type": "string",
      "title": "Part"
     },
     "has_profile": {
      "type": "boolean",
      "title": "Has Profile",
      "default": false
     },
     "entry_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entry Ids"
     }
    },
    "type": "object",
    "required": [
     "role",
     "part",
     "entry_ids"
    ],
    "title": "EciSelectionMember"
   },
...
   "EciSelectionSearch": {
    "properties": {
     "by": {
      "type": "string",
      "title": "By"
     },
     "chair_slug": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Chair Slug"
     },
     "shortlist_size": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shortlist Size"
     },
     "shortlist": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shortlist"
     },
     "shortlist_source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shortlist Source"
     },
     "entry_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entry Ids"
     }
    },
    "type": "object",
    "required": [
     "by",
     "entry_ids"
    ],
    "title": "EciSelectionSearch"
   },
...
   "EciStageValue": {
    "properties": {
     "stage": {
      "type": "string",
      "title": "Stage"
     },
     "electors": {
      "type": "integer",
      "title": "Electors"
     },
     "as_of": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of"
     },
     "computed": {
      "type": "boolean",
      "title": "Computed"
     },
     "approx": {
      "type": "boolean",
      "title": "Approx"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "source_entry_id": {
      "type": "string",
      "title": "Source Entry Id"
     },
     "source_entry_title": {
      "type": "string",
      "title": "Source Entry Title"
     },
     "source_status": {
      "type": "string",
      "title": "Source Status"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "tier": {
      "type": "integer",
      "title": "Tier"
     }
    },
    "type": "object",
    "required": [
     "stage",
     "electors",
     "computed",
     "approx",
     "source_entry_id",
     "source_entry_title",
     "source_status",
     "url",
     "tier"
    ],
    "title": "EciStageValue"
   },
   "EciStateCount": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "name",
     "count"
    ],
    "title": "EciStateCount"
   },
...
   "EciStateMetrics": {
    "properties": {
     "draft_left_off": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciStateMetric"
       },
       {
        "type": "null"
       }
      ]
     },
     "net_change": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciStateMetric"
       },
       {
        "type": "null"
       }
      ]
     },
     "appeals_filed": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EciStateMetric"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "title": "EciStateMetrics"
   },
   "EciStatePage": {
    "properties": {
     "region": {
      "$ref": "#/components/schemas/EciRegionSummary"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "region"
    ],
    "title": "EciStatePage"
   },
...
   "EciStatusCounts": {
    "properties": {
     "documented": {
      "type": "integer",
      "title": "Documented",
      "default": 0
     },
     "reported": {
      "type": "integer",
      "title": "Reported",
      "default": 0
     },
     "claim": {
      "type": "integer",
      "title": "Claim",
      "default": 0
     },
     "response": {
      "type": "integer",
      "title": "Response",
      "default": 0
     }
    },
    "type": "object",
    "title": "EciStatusCounts"
   },
...
   "EciTimeline": {
    "properties": {
     "entries": {
      "items": {
       "$ref": "#/components/schemas/EciEntry"
      },
      "type": "array",
      "title": "Entries"
     },
     "topics": {
      "items": {
       "$ref": "#/components/schemas/EciTopicCount"
      },
      "type": "array",
      "title": "Topics"
     },
     "people": {
      "items": {
       "$ref": "#/components/schemas/EciPersonCount"
      },
      "type": "array",
      "title": "People"
     },
     "lanes": {
      "items": {
       "$ref": "#/components/schemas/EciLaneCount"
      },
      "type": "array",
      "title": "Lanes"
     },
     "states": {
      "items": {
       "$ref": "#/components/schemas/EciStateCount"
      },
      "type": "array",
      "title": "States",
      "default": []
     },
     "counts": {
      "$ref": "#/components/schemas/EciCheckCounts"
     }
    },
    "type": "object",
    "required": [
     "entries",
     "topics",
     "people",
     "lanes",
     "counts"
    ],
    "title": "EciTimeline",
    "description": "The filtered ECI Files timeline \u2014 entries plus facets scoped to the same filter."
   },
   "EciTimelineCompact": {
    "properties": {
     "entries": {
      "items": {
       "$ref": "#/components/schemas/EciEntryCompact"
      },
      "type": "array",
      "title": "Entries"
     },
     "topics": {
      "items": {
       "$ref": "#/components/schemas/EciTopicCount"
      },
      "type": "array",
      "title": "Topics"
     },
     "people": {
      "items": {
       "$ref": "#/components/schemas/EciPersonCount"
      },
      "type": "array",
      "title": "People"
     },
     "lanes": {
      "items": {
       "$ref": "#/components/schemas/EciLaneCount"
      },
      "type": "array",
      "title": "Lanes"
     },
     "states": {
      "items": {
       "$ref": "#/components/schemas/EciStateCount"
      },
      "type": "array",
      "title": "States",
      "default": []
     },
     "counts": {
      "$ref": "#/components/schemas/EciCheckCounts"
     }
    },
    "type": "object",
    "required": [
     "entries",
     "topics",
     "people",
     "lanes",
     "counts"
    ],
    "title": "EciTimelineCompact",
    "description": "The `fields=compact` shape of `/eci-files/timeline`: same facets, slimmer entries."
   },
   "EciTopicCount": {
    "properties": {
     "topic": {
      "type": "string",
      "title": "Topic"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     }
    },
    "type": "object",
    "required": [
     "topic",
     "count"
    ],
    "title": "EciTopicCount"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}